Add tabBar to home
