+++ lib/models/tab_info.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+class TabInfo {
+  String label;
+  Widget widget;
+  TabInfo(this.label, this.widget);
+}
