+++ lib/widgets/home/drills_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DrillsPage extends StatelessWidget {
+  const DrillsPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
